+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Houserobbery timeout per persoon geadd
+- Admin kick from vehicle optie toegevoegd
 
 ### Changed